docker_requirements.txt
anyio==4.2.0
Bottleneck==1.3.7
certifi==2024.2.2
contourpy==1.2.0
cycler==0.11.0
fonttools==4.51.0
h11==0.14.0
httpcore==1.0.2
httpx==0.27.0
idna==3.4
kiwisolver==1.4.4
logger-tt==1.7.3
logutils==0.3.5
numexpr==2.8.7
numpy==1.26.4
packaging==23.2
pandas==2.2.1
ply==3.11
psutil==5.9.0
psycopg2==2.9.9
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==21.1.1
pytz==2023.3.post1
PyYAML==6.0.1
sigfig==1.3.3
sip==6.7.12
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tornado==6.3.3
tzdata==2023.3
unicodedata2==15.1.0

yfinance~=0.2.40
requests~=2.32.3
bs4~=0.0.2
beautifulsoup4~=4.12.3
sqlalchemy~=2.0.30